+++ R/RcppExports.R
@@ -513,8 +513,12 @@ cpp_test_s2z_block_schur <- function(A, pin_start, pin_n, pin_coef, grad) {
     .Call(`_tulpa_cpp_test_s2z_block_schur`, A, pin_start, pin_n, pin_coef, grad)
 }
 
-cpp_test_mcar_prior <- function(theta_logchol, p, n, adj_rp, adj_ci, nnbr, x) {
-    .Call(`_tulpa_cpp_test_mcar_prior`, theta_logchol, p, n, adj_rp, adj_ci, nnbr, x)
+cpp_test_log_prior_icar <- function(x, n, tau, adj_rp, adj_ci, nnbr, n_components = 1L) {
+    .Call(`_tulpa_cpp_test_log_prior_icar`, x, n, tau, adj_rp, adj_ci, nnbr, n_components)
+}
+
+cpp_test_mcar_prior <- function(theta_logchol, p, n, adj_rp, adj_ci, nnbr, x, n_components = 1L) {
+    .Call(`_tulpa_cpp_test_mcar_prior`, theta_logchol, p, n, adj_rp, adj_ci, nnbr, x, n_components)
 }
 
 cpp_test_lkj_build_L <- function(raw, n) {