+++ src/RcppExports.cpp
@@ -2349,9 +2349,26 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_test_log_prior_icar
+double cpp_test_log_prior_icar(NumericVector x, int n, double tau, IntegerVector adj_rp, IntegerVector adj_ci, IntegerVector nnbr, int n_components);
+RcppExport SEXP _tulpa_cpp_test_log_prior_icar(SEXP xSEXP, SEXP nSEXP, SEXP tauSEXP, SEXP adj_rpSEXP, SEXP adj_ciSEXP, SEXP nnbrSEXP, SEXP n_componentsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type adj_rp(adj_rpSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type adj_ci(adj_ciSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type nnbr(nnbrSEXP);
+    Rcpp::traits::input_parameter< int >::type n_components(n_componentsSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_test_log_prior_icar(x, n, tau, adj_rp, adj_ci, nnbr, n_components));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_test_mcar_prior
-List cpp_test_mcar_prior(NumericVector theta_logchol, int p, int n, IntegerVector adj_rp, IntegerVector adj_ci, IntegerVector nnbr, NumericVector x);
-RcppExport SEXP _tulpa_cpp_test_mcar_prior(SEXP theta_logcholSEXP, SEXP pSEXP, SEXP nSEXP, SEXP adj_rpSEXP, SEXP adj_ciSEXP, SEXP nnbrSEXP, SEXP xSEXP) {
+List cpp_test_mcar_prior(NumericVector theta_logchol, int p, int n, IntegerVector adj_rp, IntegerVector adj_ci, IntegerVector nnbr, NumericVector x, int n_components);
+RcppExport SEXP _tulpa_cpp_test_mcar_prior(SEXP theta_logcholSEXP, SEXP pSEXP, SEXP nSEXP, SEXP adj_rpSEXP, SEXP adj_ciSEXP, SEXP nnbrSEXP, SEXP xSEXP, SEXP n_componentsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -2362,7 +2379,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< IntegerVector >::type adj_ci(adj_ciSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type nnbr(nnbrSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_test_mcar_prior(theta_logchol, p, n, adj_rp, adj_ci, nnbr, x));
+    Rcpp::traits::input_parameter< int >::type n_components(n_componentsSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_test_mcar_prior(theta_logchol, p, n, adj_rp, adj_ci, nnbr, x, n_components));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -2942,7 +2960,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_tulpa_cpp_test_spde_nc_transform_grad", (DL_FUNC) &_tulpa_cpp_test_spde_nc_transform_grad, 10},
     {"_tulpa_cpp_test_spde_nc_transform_fwd", (DL_FUNC) &_tulpa_cpp_test_spde_nc_transform_fwd, 10},
     {"_tulpa_cpp_test_s2z_block_schur", (DL_FUNC) &_tulpa_cpp_test_s2z_block_schur, 5},
-    {"_tulpa_cpp_test_mcar_prior", (DL_FUNC) &_tulpa_cpp_test_mcar_prior, 7},
+    {"_tulpa_cpp_test_log_prior_icar", (DL_FUNC) &_tulpa_cpp_test_log_prior_icar, 7},
+    {"_tulpa_cpp_test_mcar_prior", (DL_FUNC) &_tulpa_cpp_test_mcar_prior, 8},
     {"_tulpa_cpp_test_lkj_build_L", (DL_FUNC) &_tulpa_cpp_test_lkj_build_L, 2},
     {"_tulpa_cpp_test_lkj_density", (DL_FUNC) &_tulpa_cpp_test_lkj_density, 2},
     {"_tulpa_cpp_test_lkj_grad", (DL_FUNC) &_tulpa_cpp_test_lkj_grad, 3},